backend/requirements.txt
# EduPlatform Backend - Full Requirements
# Install: pip install -r requirements.txt

# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
starlette==0.41.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.2.0  # Required for Pydantic email validation

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.6
greenlet==3.2.3  # Required for SQLAlchemy async operations

# Redis & Caching
redis==5.2.1

# Celery & Task Queue
celery==5.4.0
flower==2.0.1

# OpenAI & AI
openai==1.57.2
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.63
langchain-openai==0.2.12
tiktoken==0.12.0

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.3.0

# File Processing
pdfplumber==0.11.4
PyMuPDF==1.25.1
pypdfium2==5.1.0
python-docx==1.1.2
mammoth==1.8.0
odfpy==1.4.1
ebooklib==0.18
markdown==3.7
striprtf==0.0.26
beautifulsoup4==4.12.3  # Required for HTML parsing
lxml==5.3.0  # Required for document processing

# YouTube & Media
yt-dlp==2025.1.26
youtube-transcript-api==0.6.3
edge-tts==7.2.7
pydub==0.25.1
audioop-lts==0.2.2

# Web Search
tavily-python==0.5.0

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.23.1

# Retry Logic
tenacity==9.1.2

# HTTP Client
httpx==0.28.1
httpx-sse==0.4.3

# Utilities
python-multipart==0.0.20
python-slugify==8.0.4
orjson==3.11.4
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Development
flake8==7.1.1
black==24.10.0
mypy==1.13.0
ipdb==0.13.13
